Update start.py
Updated the json format.
Added the http-server.
Requests to http://localhost:5000/dynamic will return the dynamic data and requests to http://localhost:5000/static will return the static data. A json post to http://localhost:5000/slider with, e.g., {"name": "foo", "value": "0.1"} will adjust the slider with the respective name if the value is within the bounds.
Additionally, Requests to http://localhost:5000/pause will pause (only!) the KMC simulation in the background and Requests to http://localhost:5000/resume will resume it.